Fix calculation of Michigan retirement deduction calculation Fixes #3456
