+++ policyengine_us/tests/policy/baseline/gov/states/mi/tax/income/deductions/retirement/tier_three/mi_retirement_benefits_deduction_tier_three.yaml
@@ -3,10 +3,10 @@
   input:
     mi_retirement_benefits_deduction_tier_three_eligible: true
     mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired: 15_000
-    mi_retirement_benefits_deduction_tier_three_ss_exempt_retired: 40_000
+    mi_retirement_benefits_deduction_tier_three_ss_exempt_retired: 35_000
     state_code: MI
-  output: # 40000 > 15000
-    mi_retirement_benefits_deduction_tier_three: 40_000
+  output: # 35000 > 15000
+    mi_retirement_benefits_deduction_tier_three: 35_000
 
 - name: 2022 filer with tier 3 retirement and pension benefits, ss_exempt_retired smaller than ss_exempt_not_retired amount
   period: 2022
