+++ policyengine_us/tests/policy/baseline/gov/states/mi/tax/income/deductions/retirement/tier_three/ss_exempt/not_retired/mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired.yaml
@@ -4,17 +4,19 @@
     people:
       person1:
         taxable_pension_income: 40_000
+        military_retirement_pay: 1_000
         is_tax_unit_head_or_spouse: true
     tax_units:
       tax_unit:
         mi_retirement_benefits_deduction_tier_three_eligible: true
         mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired_eligible_people: 1
+        mi_retirement_benefits_deduction_tier_one_amount: 40_000
         members: [person1]
         filing_status: SINGLE
     household:
       members: [person1]
       state_code: MI
-  output: # 0 + 0 + 15000 + 0 < 40000
+  output: # min(0 + 0 + 15000 + 0 < 40000, 40000)
     mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired: 15_000
 
 - name: 2022 joint filer and their spouse with full tier 3 retirement and pension benefits larger than taxable pension income, and with one more household member
@@ -23,6 +25,7 @@
     people:
       person1:
         taxable_pension_income: 10_000
+        military_retirement_pay: 1_000
         is_tax_unit_head_or_spouse: true
       person2:
         taxable_pension_income: 10_000
@@ -34,10 +37,32 @@
       tax_unit:
         mi_retirement_benefits_deduction_tier_three_eligible: true
         mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired_eligible_people: 2
+        mi_retirement_benefits_deduction_tier_one_amount: 30_000
         members: [person1, person2, person3]
         filing_status: JOINT
     household:
       members: [person1, person2, person3]
       state_code: MI
-  output: # 15000 + 15000 > 10000 + 10000
+  output: # min(15000 + 15000 > 10000 + 10000, 30000)
     mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired: 20_000
+
+- name: 2022 single filer with full tier 3 retirement and pension benefits larger than tier 1 retirement and pension benefits
+  period: 2022
+  input:
+    people:
+      person1:
+        taxable_pension_income: 40_000
+        military_retirement_pay: 1_000
+        is_tax_unit_head_or_spouse: true
+    tax_units:
+      tax_unit:
+        mi_retirement_benefits_deduction_tier_three_eligible: true
+        mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired_eligible_people: 1
+        mi_retirement_benefits_deduction_tier_one_amount: 10_000
+        members: [person1]
+        filing_status: SINGLE
+    household:
+      members: [person1]
+      state_code: MI
+  output: # min(0 + 0 + 15000 + 0 < 40000, 10000)
+    mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired: 10_000
