+++ policyengine_us/tests/policy/baseline/gov/states/mi/tax/income/deductions/retirement/tier_three/ss_exempt/retired/mi_retirement_benefits_deduction_tier_three_ss_exempt_retired.yaml
@@ -4,6 +4,7 @@
     people:
       person1:
         taxable_pension_income: 20_000
+        military_retirement_pay: 1_000
         is_tax_unit_head_or_spouse: true
       person2:
         taxable_pension_income: 40_000
@@ -12,12 +13,13 @@
       tax_unit:
         mi_retirement_benefits_deduction_tier_three_eligible: true
         mi_retirement_benefits_deduction_tier_three_ss_exempt_retired_eligible_people: 2
+        mi_retirement_benefits_deduction_tier_one_amount: 80_000
         members: [person1, person2]
         filing_status: JOINT
     household:
       members: [person1, person2]
       state_code: MI
-  output: # 80000 > 20000 + 40000
+  output: # min(70000 > 20000 + 40000, 80000)
     mi_retirement_benefits_deduction_tier_three_ss_exempt_retired: 60_000
 
 - name: 2022 joint filer and their spouse with tier 3 retirement and pension benefits, one receive retirement benefits from SSA employment exempt and retired before 2013, smaller than taxable pension income, and with one more household member
@@ -26,6 +28,7 @@
     people:
       person1:
         taxable_pension_income: 40_000
+        military_retirement_pay: 1_000
         is_tax_unit_head_or_spouse: true
       person2:
         taxable_pension_income: 30_000
@@ -37,10 +40,32 @@
       tax_unit:
         mi_retirement_benefits_deduction_tier_three_eligible: true
         mi_retirement_benefits_deduction_tier_three_ss_exempt_retired_eligible_people: 1
+        mi_retirement_benefits_deduction_tier_one_amount: 60_000
         members: [person1, person2, person3]
         filing_status: JOINT
     household:
       members: [person1, person2, person3]
       state_code: MI
-  output: # 55000 < 40000 + 30000
+  output: # min(55000 < 40000 + 30000, 60000)
     mi_retirement_benefits_deduction_tier_three_ss_exempt_retired: 55_000
+
+- name: 2022 single filer with tier 3 retirement and pension benefits, receive retirement benefits from SSA employment exempt and retired before 2013, larger than tier 1 retirement and pension benefits
+  period: 2022
+  input:
+    people:
+      person1:
+        taxable_pension_income: 40_000
+        military_retirement_pay: 1_000
+        is_tax_unit_head_or_spouse: true
+    tax_units:
+      tax_unit:
+        mi_retirement_benefits_deduction_tier_three_eligible: true
+        mi_retirement_benefits_deduction_tier_three_ss_exempt_retired_eligible_people: 1
+        mi_retirement_benefits_deduction_tier_one_amount: 10_000
+        members: [person1]
+        filing_status: SINGLE
+    household:
+      members: [person1]
+      state_code: MI
+  output: # min(35000 < 40000, 10000)
+    mi_retirement_benefits_deduction_tier_three_ss_exempt_retired: 10_000
