+++ policyengine_us/variables/gov/states/mi/tax/income/deductions/retirement/tier_three/ss_exempt/not_retired/mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired.py
@@ -13,6 +13,7 @@ class mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired(
         "http://legislature.mi.gov/doc.aspx?mcl-206-30",  # (9)(d)
         "https://www.michigan.gov/taxes/-/media/Project/Websites/taxes/Forms/2022/2022-IIT-Forms/BOOK_MI-1040.pdf#page=17",
         "https://www.michigan.gov/taxes/iit/retirement-and-pension-benefits",
+        "https://www.michigan.gov/taxes/-/media/Project/Websites/taxes/Forms/2022/2022-IIT-Forms/Form-4884-Section-C-worksheet.pdf",
     )
     defined_for = "mi_retirement_benefits_deduction_tier_three_eligible"
 
@@ -33,6 +34,21 @@ class mi_retirement_benefits_deduction_tier_three_ss_exempt_not_retired(
         # Head and spouse both are eligible to receive an equal deduction amount
         cap = p.amount * eligible_people
 
-        return min_(
+        base_amount = min_(
             tax_unit.sum(uncapped_pension_income * is_head_or_spouse), cap
+        )  # Line 9
+
+        military_retirement_pay_eligible = (
+            tax_unit.sum(person("military_retirement_pay", period)) > 0
         )
+
+        tier_one_amount = tax_unit(
+            "mi_retirement_benefits_deduction_tier_one_amount",
+            period,
+        )  # Line 8
+
+        return where(
+            military_retirement_pay_eligible,
+            min_(base_amount, tier_one_amount),
+            base_amount,
+        )  # Line 10
